Se guarda la posición absoluta de los motores como la posición de los brazos.

+++ D2/DeltaRobot.h
@@ -9,7 +9,7 @@ class DeltaRobot {
     DeltaRobot(){};
     typedef struct 
     {
-      float position;
+      long position;
       byte homed;
     } arm;
